Rotation and Tetromino Numbers
Update tile class by import random for randomly select 2 and 4 for fill tetrominos. Then update tetromino class, define rotate movement next to the move left and rigt (the other movements) then set the criterias for rotation because when we on the right or left most of the grid when the tetromino rotated out of the frame and  directl game end.